mvp_agente_logistica/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# OCR
paddleocr==2.7.3
paddlepaddle==2.6.0
pytesseract==0.3.10

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
pdf2image==1.16.3
numpy==1.26.2

# LLM
ollama==0.1.6
requests==2.31.0

# CLI
typer[all]==0.9.0
rich==13.7.0
click==8.1.7

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
tqdm==4.66.1
tenacity==8.2.3
